requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.2.1=pypi_0
attrs=22.2.0=pypi_0
backcall=0.2.0=pypi_0
bcrypt=4.0.1=pypi_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.0.1=pypi_0
comm=0.1.2=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.2=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.6=pypi_0
distro=1.8.0=pypi_0
docker=6.0.1=pypi_0
docker-compose=1.29.2=pypi_0
dockerpty=0.4.1=pypi_0
docopt=0.6.2=pypi_0
executing=1.2.0=pypi_0
fonttools=4.38.0=pypi_0
google-pasta=0.2.0=pypi_0
importlib-metadata=4.13.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipykernel=6.21.2=pypi_0
ipython=8.11.0=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=8.0.3=pypi_0
jupyter-core=5.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
numpy=1.24.2=pypi_0
openssl=1.1.1t=h7f8727e_0
pandas=1.5.3=pypi_0
paramiko=3.0.0=pypi_0
parso=0.8.3=pypi_0
pathos=0.3.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py39h06a4308_0
platformdirs=3.1.0=pypi_0
pox=0.3.2=pypi_0
ppft=1.7.6.6=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
protobuf3-to-dict=0.1.5=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pypi_0
pygments=2.14.0=pypi_0
pyinstrument=3.4.2=pypi_0
pyinstrument-cext=0.2.4=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.9.16=h7a1cb2a_1
python-dotenv=0.21.1=pypi_0
pytz=2022.7.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=25.0.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
sagemaker=2.135.1.post0=pypi_0
schema=0.7.5=pypi_0
scikit-learn=1.2.1=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pypi_0
smdebug=1.0.12=pypi_0
smdebug-rulesconfig=1.0.1=pypi_0
sqlite=3.40.1=h5082296_0
stack-data=0.6.2=pypi_0
texttable=1.6.7=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.13.1+cu116=pypi_0
torchvision=0.14.1+cu116=pypi_0
tornado=6.2=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022g=h04d1e81_0
wcwidth=0.2.6=pypi_0
websocket-client=0.59.0=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.2.10=h5eee18b_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
